my-app/dataspell/requirements.txt
# Python package requirements for DataSpell PostgreSQL integration

# Database connectivity
psycopg2-binary>=2.9.3
SQLAlchemy>=1.4.0

# Data analysis and visualization
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Utilities
python-dotenv>=0.19.0
faker>=8.0.0  # For generating sample data
